Moving ipa malicious test into oneshot. Previous test did not validate the results were same as expected.
